Build RiveSharp with clang
We may want to try and standardize on clang SIMD vector extensions,
which means everything has to build with clang.

Diffs=
07186fc98 Build RiveSharp with clang (#4194)

+++ include/rive/math/contour_measure.hpp
@@ -15,7 +15,7 @@ namespace rive {
 class ContourMeasure : public RefCnt {
 public:
     static constexpr unsigned kMaxDot30 = (1 << 30) - 1;
-    static constexpr float kInvScaleD30 = 1.0f / kMaxDot30;
+    static constexpr float kInvScaleD30 = 1.0f / (float)kMaxDot30;
 
     // Deliberately making this pack well (12 bytes)
     struct Segment {
